update: move MatrixKind and fix tests for storage.utils
