+++ junifer/storage/__init__.pyi
@@ -8,7 +8,8 @@ __all__ = [
     "Upsert",
 ]
 
-from .base import BaseFeatureStorage, MatrixKind, StorageType
+from ._types import MatrixKind
+from .base import BaseFeatureStorage, StorageType
 from .hdf5 import HDF5FeatureStorage
 from .pandas_base import PandasBaseFeatureStorage
 from .sqlite import SQLiteFeatureStorage, Upsert
